services/users/requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.4
psycopg2-binary
Flask-Testing==0.7.1
gunicorn==19.7.1
coverage==4.4.2
flake8===3.5.0
flask-debugtoolbar==0.10.1
bcrypt==3.1.4
Flask-Migrate